Fix circuit behavior for non-consecutive failures.
This fixes handling for non-consecutive failures, which should not cause
the circuit to open. Upon every successul block execution, it now resets
the failure counter to 0. That also means that the "half_open" state
becomes completely unneeded and the "tripped?" check becomes much
simpler and easier to understand as a result.